+++ internal/xsql/conn.go
@@ -73,7 +73,7 @@ type conn struct {
 }
 
 func (c *conn) IsValid() bool {
-	return !c.isClosed()
+	return c.isReady()
 }
 
 type currentTx interface {
@@ -91,7 +91,6 @@ var (
 	_ driver.QueryerContext     = &conn{}
 	_ driver.Pinger             = &conn{}
 	_ driver.NamedValueChecker  = &conn{}
-	_ driver.SessionResetter    = &conn{}
 	_ driver.Validator          = &conn{}
 )
 
@@ -103,25 +102,12 @@ func newConn(c *Connector, s table.ClosableSession, opts ...connOption) *conn {
 	for _, o := range opts {
 		o(cc)
 	}
+	c.attach(cc)
 	return cc
 }
 
-func (c *conn) checkClosed(err error) error {
-	if err = badconn.Map(err); xerrors.Is(err, driver.ErrBadConn) {
-		atomic.StoreUint32(&c.closed, 1)
-	}
-	return err
-}
-
-func (c *conn) isClosed() bool {
-	if atomic.LoadUint32(&c.closed) == 1 {
-		return true
-	}
-	if c.session.Status() != table.SessionReady {
-		atomic.StoreUint32(&c.closed, 1)
-		return true
-	}
-	return false
+func (c *conn) isReady() bool {
+	return c.session.Status() == table.SessionReady
 }
 
 func (conn) CheckNamedValue(v *driver.NamedValue) (err error) {
@@ -133,14 +119,18 @@ func (c *conn) PrepareContext(ctx context.Context, query string) (_ driver.Stmt,
 	defer func() {
 		onDone(err)
 	}()
-	if c.isClosed() {
-		return nil, errClosedConn
+	if !c.isReady() {
+		return nil, badconn.Map(xerrors.WithStackTrace(errNotReadyConn))
 	}
 	return &stmt{
 		conn:  c,
 		query: query,
 		trace: c.trace,
 	}, nil
+}
+
+func (c *conn) sinceLastUsage() time.Duration {
+	return time.Since(time.Unix(atomic.LoadInt64(&c.lastUsage), 0))
 }
 
 func (c *conn) execContext(ctx context.Context, query string, args []driver.NamedValue) (_ driver.Result, err error) {
@@ -151,7 +141,7 @@ func (c *conn) execContext(ctx context.Context, query string, args []driver.Name
 		query,
 		m.String(),
 		xcontext.IsIdempotent(ctx),
-		time.Since(time.Unix(atomic.LoadInt64(&c.lastUsage), 0)),
+		c.sinceLastUsage(),
 	)
 	defer func() {
 		atomic.StoreInt64(&c.lastUsage, time.Now().Unix())
@@ -167,38 +157,38 @@ func (c *conn) execContext(ctx context.Context, query string, args []driver.Name
 			dataQueryOptions(ctx)...,
 		)
 		if err != nil {
-			return nil, c.checkClosed(xerrors.WithStackTrace(err))
+			return nil, badconn.Map(xerrors.WithStackTrace(err))
 		}
 		defer func() {
 			_ = res.Close()
 		}()
 		if err = res.NextResultSetErr(ctx); !xerrors.Is(err, nil, io.EOF) {
-			return nil, c.checkClosed(xerrors.WithStackTrace(err))
+			return nil, badconn.Map(xerrors.WithStackTrace(err))
 		}
 		if err = res.Err(); err != nil {
-			return nil, c.checkClosed(xerrors.WithStackTrace(err))
+			return nil, badconn.Map(xerrors.WithStackTrace(err))
 		}
 		return driver.ResultNoRows, nil
 	case SchemeQueryMode:
 		err = c.session.ExecuteSchemeQuery(ctx, query)
 		if err != nil {
-			return nil, c.checkClosed(xerrors.WithStackTrace(err))
+			return nil, badconn.Map(xerrors.WithStackTrace(err))
 		}
 		return driver.ResultNoRows, nil
 	case ScriptingQueryMode:
 		var res result.StreamResult
 		res, err = c.connector.connection.Scripting().StreamExecute(ctx, query, toQueryParams(args))
 		if err != nil {
-			return nil, c.checkClosed(xerrors.WithStackTrace(err))
+			return nil, badconn.Map(xerrors.WithStackTrace(err))
 		}
 		defer func() {
 			_ = res.Close()
 		}()
 		if err = res.NextResultSetErr(ctx); !xerrors.Is(err, nil, io.EOF) {
-			return nil, c.checkClosed(xerrors.WithStackTrace(err))
+			return nil, badconn.Map(xerrors.WithStackTrace(err))
 		}
 		if err = res.Err(); err != nil {
-			return nil, c.checkClosed(xerrors.WithStackTrace(err))
+			return nil, badconn.Map(xerrors.WithStackTrace(err))
 		}
 		return driver.ResultNoRows, nil
 	default:
@@ -207,8 +197,8 @@ func (c *conn) execContext(ctx context.Context, query string, args []driver.Name
 }
 
 func (c *conn) ExecContext(ctx context.Context, query string, args []driver.NamedValue) (_ driver.Result, err error) {
-	if c.isClosed() {
-		return nil, errClosedConn
+	if !c.isReady() {
+		return nil, badconn.Map(xerrors.WithStackTrace(errNotReadyConn))
 	}
 	if c.currentTx != nil {
 		return c.currentTx.ExecContext(ctx, query, args)
@@ -217,8 +207,8 @@ func (c *conn) ExecContext(ctx context.Context, query string, args []driver.Name
 }
 
 func (c *conn) QueryContext(ctx context.Context, query string, args []driver.NamedValue) (_ driver.Rows, err error) {
-	if c.isClosed() {
-		return nil, errClosedConn
+	if !c.isReady() {
+		return nil, badconn.Map(xerrors.WithStackTrace(errNotReadyConn))
 	}
 	if c.currentTx != nil {
 		return c.currentTx.QueryContext(ctx, query, args)
@@ -234,7 +224,7 @@ func (c *conn) queryContext(ctx context.Context, query string, args []driver.Nam
 		query,
 		m.String(),
 		xcontext.IsIdempotent(ctx),
-		time.Since(time.Unix(atomic.LoadInt64(&c.lastUsage), 0)),
+		c.sinceLastUsage(),
 	)
 	defer func() {
 		atomic.StoreInt64(&c.lastUsage, time.Now().Unix())
@@ -250,10 +240,10 @@ func (c *conn) queryContext(ctx context.Context, query string, args []driver.Nam
 			dataQueryOptions(ctx)...,
 		)
 		if err != nil {
-			return nil, c.checkClosed(xerrors.WithStackTrace(err))
+			return nil, badconn.Map(xerrors.WithStackTrace(err))
 		}
 		if err = res.Err(); err != nil {
-			return nil, c.checkClosed(xerrors.WithStackTrace(err))
+			return nil, badconn.Map(xerrors.WithStackTrace(err))
 		}
 		return &rows{
 			conn:   c,
@@ -267,10 +257,10 @@ func (c *conn) queryContext(ctx context.Context, query string, args []driver.Nam
 			scanQueryOptions(ctx)...,
 		)
 		if err != nil {
-			return nil, c.checkClosed(xerrors.WithStackTrace(err))
+			return nil, badconn.Map(xerrors.WithStackTrace(err))
 		}
 		if err = res.Err(); err != nil {
-			return nil, c.checkClosed(xerrors.WithStackTrace(err))
+			return nil, badconn.Map(xerrors.WithStackTrace(err))
 		}
 		return &rows{
 			conn:   c,
@@ -280,7 +270,7 @@ func (c *conn) queryContext(ctx context.Context, query string, args []driver.Nam
 		var exp table.DataQueryExplanation
 		exp, err = c.session.Explain(ctx, query)
 		if err != nil {
-			return nil, c.checkClosed(xerrors.WithStackTrace(err))
+			return nil, badconn.Map(xerrors.WithStackTrace(err))
 		}
 		return &single{
 			values: []sql.NamedArg{
@@ -292,10 +282,10 @@ func (c *conn) queryContext(ctx context.Context, query string, args []driver.Nam
 		var res result.StreamResult
 		res, err = c.connector.connection.Scripting().StreamExecute(ctx, query, toQueryParams(args))
 		if err != nil {
-			return nil, c.checkClosed(xerrors.WithStackTrace(err))
+			return nil, badconn.Map(xerrors.WithStackTrace(err))
 		}
 		if err = res.Err(); err != nil {
-			return nil, c.checkClosed(xerrors.WithStackTrace(err))
+			return nil, badconn.Map(xerrors.WithStackTrace(err))
 		}
 		return &rows{
 			conn:   c,
@@ -311,25 +301,29 @@ func (c *conn) Ping(ctx context.Context) (err error) {
 	defer func() {
 		onDone(err)
 	}()
-	if c.isClosed() {
-		return errClosedConn
+	if !c.isReady() {
+		return badconn.Map(xerrors.WithStackTrace(errNotReadyConn))
 	}
 	if err = c.session.KeepAlive(ctx); err != nil {
-		return c.checkClosed(xerrors.WithStackTrace(err))
+		return badconn.Map(xerrors.WithStackTrace(err))
 	}
 	return nil
 }
 
 func (c *conn) Close() (err error) {
-	onDone := trace.DatabaseSQLOnConnClose(c.trace)
-	defer func() {
-		onDone(err)
-	}()
-	err = c.session.Close(context.Background())
-	if err != nil {
-		return c.checkClosed(xerrors.WithStackTrace(err))
+	if atomic.CompareAndSwapUint32(&c.closed, 0, 1) {
+		c.connector.detach(c)
+		onDone := trace.DatabaseSQLOnConnClose(c.trace)
+		defer func() {
+			onDone(err)
+		}()
+		err = c.session.Close(context.Background())
+		if err != nil {
+			return badconn.Map(xerrors.WithStackTrace(err))
+		}
+		return nil
 	}
-	return nil
+	return badconn.Map(xerrors.WithStackTrace(errConnClosedEarly))
 }
 
 func (c *conn) Prepare(string) (driver.Stmt, error) {
@@ -346,8 +340,8 @@ func (c *conn) BeginTx(ctx context.Context, txOptions driver.TxOptions) (_ drive
 	defer func() {
 		onDone(transaction, err)
 	}()
-	if c.isClosed() {
-		return nil, errClosedConn
+	if !c.isReady() {
+		return nil, badconn.Map(xerrors.WithStackTrace(errNotReadyConn))
 	}
 	if c.currentTx != nil {
 		return nil, xerrors.WithStackTrace(
@@ -361,7 +355,7 @@ func (c *conn) BeginTx(ctx context.Context, txOptions driver.TxOptions) (_ drive
 	}
 	transaction, err = c.session.BeginTransaction(ctx, table.TxSettings(txc))
 	if err != nil {
-		return nil, c.checkClosed(xerrors.WithStackTrace(err))
+		return nil, badconn.Map(xerrors.WithStackTrace(err))
 	}
 	c.currentTx = &tx{
 		conn: c,
@@ -369,16 +363,6 @@ func (c *conn) BeginTx(ctx context.Context, txOptions driver.TxOptions) (_ drive
 		tx:   transaction,
 	}
 	return c.currentTx, nil
-}
-
-func (c *conn) ResetSession(_ context.Context) error {
-	if c.currentTx != nil {
-		_ = c.currentTx.Rollback()
-	}
-	if c.isClosed() {
-		return errClosedConn
-	}
-	return nil
 }
 
 func (c *conn) Version(ctx context.Context) (_ string, err error) {