+++ internal/xsql/connector.go
@@ -4,6 +4,8 @@ import (
 	"context"
 	"database/sql/driver"
 	"io"
+	"sync"
+	"time"
 
 	metaHeaders "github.com/ydb-platform/ydb-go-sdk/v3/internal/meta"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
@@ -52,10 +54,25 @@ func WithTrace(t trace.DatabaseSQL, opts ...trace.DatabaseSQLComposeOption) Conn
 	}
 }
 
+func WithDisableServerBalancer() ConnectorOption {
+	return func(c *Connector) error {
+		c.disableServerBalancer = true
+		return nil
+	}
+}
+
+func WithIdleThreshold(idleThreshold time.Duration) ConnectorOption {
+	return func(c *Connector) error {
+		c.idleThreshold = idleThreshold
+		return nil
+	}
+}
+
 func Open(d Driver, connection connection, opts ...ConnectorOption) (_ *Connector, err error) {
 	c := &Connector{
 		driver:           d,
 		connection:       connection,
+		conns:            make(map[*conn]struct{}),
 		defaultTxControl: table.DefaultTxControl(),
 		defaultQueryMode: DefaultQueryMode,
 	}
@@ -63,6 +80,9 @@ func Open(d Driver, connection connection, opts ...ConnectorOption) (_ *Connecto
 		if err = opt(c); err != nil {
 			return nil, err
 		}
+	}
+	if c.idleThreshold > 0 {
+		c.idleStopper = c.idleCloser()
 	}
 	d.Attach(c)
 	return c, nil
@@ -94,10 +114,17 @@ type Connector struct {
 	driver     Driver
 	connection connection
 
-	defaultTxControl     *table.TransactionControl
-	defaultQueryMode     QueryMode
-	defaultDataQueryOpts []options.ExecuteDataQueryOption
-	defaultScanQueryOpts []options.ExecuteScanQueryOption
+	conns    map[*conn]struct{}
+	connsMtx sync.RWMutex
+
+	idleStopper func()
+
+	defaultTxControl      *table.TransactionControl
+	defaultQueryMode      QueryMode
+	defaultDataQueryOpts  []options.ExecuteDataQueryOption
+	defaultScanQueryOpts  []options.ExecuteScanQueryOption
+	disableServerBalancer bool
+	idleThreshold         time.Duration
 
 	trace trace.DatabaseSQL
 }
@@ -107,8 +134,37 @@ var (
 	_ io.Closer        = &Connector{}
 )
 
+func (c *Connector) idleCloser() (idleStopper func()) {
+	var ctx context.Context
+	ctx, idleStopper = context.WithCancel(context.Background())
+	go func() {
+		for {
+			select {
+			case <-ctx.Done():
+				return
+			case <-time.After(c.idleThreshold):
+				c.connsMtx.RLock()
+				conns := make([]*conn, 0, len(c.conns))
+				for cc := range c.conns {
+					conns = append(conns, cc)
+				}
+				c.connsMtx.RUnlock()
+				for _, cc := range conns {
+					if cc.sinceLastUsage() > c.idleThreshold {
+						cc.session.Close(context.Background())
+					}
+				}
+			}
+		}
+	}()
+	return idleStopper
+}
+
 func (c *Connector) Close() (err error) {
 	defer c.driver.Detach(c)
+	if c.idleStopper != nil {
+		c.idleStopper()
+	}
 	return nil
 }
 
@@ -116,16 +172,27 @@ func (c *Connector) Connection() connection {
 	return c.connection
 }
 
+func (c *Connector) attach(cc *conn) {
+	c.connsMtx.Lock()
+	defer c.connsMtx.Unlock()
+	c.conns[cc] = struct{}{}
+}
+
+func (c *Connector) detach(cc *conn) {
+	c.connsMtx.Lock()
+	defer c.connsMtx.Unlock()
+	delete(c.conns, cc)
+}
+
 func (c *Connector) Connect(ctx context.Context) (_ driver.Conn, err error) {
 	onDone := trace.DatabaseSQLOnConnectorConnect(c.trace, &ctx)
 	defer func() {
 		onDone(err)
 	}()
-	s, err := c.connection.Table().CreateSession( //nolint:staticcheck // SA1019
-		meta.WithAllowFeatures(ctx,
-			metaHeaders.HintSessionBalancer,
-		),
-	)
+	if !c.disableServerBalancer {
+		ctx = meta.WithAllowFeatures(ctx, metaHeaders.HintSessionBalancer)
+	}
+	s, err := c.connection.Table().CreateSession(ctx) //nolint:staticcheck // SA1019
 	if err != nil {
 		return nil, xerrors.WithStackTrace(err)
 	}