+++ internal/xsql/unwrap.go
@@ -16,7 +16,7 @@ func Unwrap(db *sql.DB) (connector *Connector, err error) {
 	// hop with create session (connector.Connect()) helps to get ydb.Connection
 	c, err := db.Conn(context.Background())
 	if err != nil {
-		return nil, xerrors.WithStackTrace(err)
+		return nil, badconn.Map(xerrors.WithStackTrace(err))
 	}
 	if err = c.Raw(func(driverConn interface{}) error {
 		if cc, ok := driverConn.(*conn); ok {
@@ -25,7 +25,7 @@ func Unwrap(db *sql.DB) (connector *Connector, err error) {
 		}
 		return xerrors.WithStackTrace(badconn.Map(fmt.Errorf("%+v is not a *conn", driverConn)))
 	}); err != nil {
-		return nil, xerrors.WithStackTrace(err)
+		return nil, badconn.Map(xerrors.WithStackTrace(err))
 	}
 	return connector, nil
 }