+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+* Fixed closing of `database/sql` connection (aka `YDB` session)
+* Made `session.Close()` as `nop` for idled session
+* Implemented goroutine for closing idle connection in `database/sql` driver
+* Separated errors of commit from other reader and to expired session
+* Fixed wrapping error in `internal/balancer/Balancer.wrapCall()`
+
+## v3.42.4
+* Added `ydb.WithDisableServerBalancer()` database/sql connector option
+
+## v3.42.3
+* Added `credentials.NewStaticCredentials()` static credentials constructor
+* Changed `internal/credentials.NewStaticCredentials()` signature and behaviour for create grpc connection on each call to auth service 
+* Downgrade `google.golang.org/grpc` to `v1.49.0`
+
+## v3.42.2
+* Added `trace.Details.Details()` method for use external detailer
+
 ## v3.42.1
 * Fixed lazy transaction example for `godoc`